Bluetooth: HID: Dont send hid connect after sdp complete
This patch prevents btif layer from calling btif_hh_connect immediately
after SDP is completed for HID Pointing device.  Without this patch it
was seen that hid connection was initiated from upper layers after sdp
information was processed, which was leading to error in logs that hid
connection was not successful.

Change-Id: I32940fed9c8ed268d6f7af00c86fb238619e0fb7
CRs-Fixed: 582935